FIXED BURRY BUG AND ADDED DESK TESTS
I've fixed the issue with sprites being 'blurred' or otherwise just
being un-sharp when imported into Unity. Buried inside this changelog is
a document detailing how to make sure sprites won't be blurry in the off
chance I don't already go through the motions on my end.

In addition, I've created two sprites of a desk, more so as a test to
make sure importing sprites worked than anything else (though we can use
it as decoration for the facility in the levels). There are two versions
that I made: a 32x32 version and a 64x64 version (appropriately named
Desk Test - 32x32 and Desk Test - 64x64). The 64 version looks much
better than the 32 one, but takes a SIGNIFICANTLY longer amount of time
to create. Personally, I think it would be worth going with 64 anyways,
even though it would take much more time.